LayoutTests/http/tests/referrer-policy-iframe/same-origin/cross-origin-http-http.html

<!DOCTYPE html>
<html>
<head>
<script src="/js-test-resources/js-test.js"></script>
</head>
<body>
<script>
description("Tests the behavior of same-origin referrer policy when cross origin.");
jsTestIsAsync = true;

if (window.internals)
    internals.setTrackingPreventionEnabled(false);

window.onmessage = function(event) {
    referrer = event.data.referrer;
    // Should be the empty string because we are cross-origin.
    shouldBeEqualToString("referrer", "");
    if (window.internals)
        internals.setTrackingPreventionEnabled(true);
    finishJSTest();
}
</script>
<iframe src="http://localhost:8000/referrer-policy/resources/document.html" referrerpolicy="same-origin"></iframe>
</body>
</html>
