LayoutTests/http/tests/referrer-policy-iframe/same-origin/cross-origin-http-http-expected.txt

Tests the behavior of same-origin referrer policy when cross origin.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
